relocate function `single_experiment`
