+++ experiments/estimate_ate/hr_estimation.py
@@ -7,8 +7,9 @@ import numpy as np
 import pandas as pd
 import yaml
 
+from experiments.synthetic import single_experiment
 from fedeca.competitors import MatchingAjudsted, NaiveComparison, PooledIPTW
-from fedeca.utils.experiment_utils import param_grid_from_dict, single_experiment
+from fedeca.utils.experiment_utils import param_grid_from_dict
 from fedeca.utils.survival_utils import CoxData
 
 if __name__ == "__main__":
